Final Commit

+++ backend/controllers/EnrollmentController.js
@@ -1,4 +1,5 @@
 const EnrollmentService = require('../services/EnrollmentService');
+const CourseService = require('../services/CourseService');
 const UserService = require('../services/UserService');
 const { sendSuccess, sendError } = require('../utils/response');
 
@@ -21,12 +22,27 @@ exports.create = async (req, res, next) => {
 
 exports.getAll = async (req, res, next) => {
   try {
-    const data = req.user.role === 'admin'
-      ? await EnrollmentService.getAll()
-      : req.user.role === 'learner'
-        ? await EnrollmentService.getByUserId(req.user.sub)
-        : [];
-    sendSuccess(res, 200, data);
+    if (req.user.role === 'admin') {
+      const data = await EnrollmentService.getAll();
+      return sendSuccess(res, 200, data);
+    }
+
+    if (req.user.role === 'learner') {
+      const data = await EnrollmentService.getByUserId(req.user.sub);
+      return sendSuccess(res, 200, data);
+    }
+
+    if (req.user.role === 'instructor') {
+      const courses = await CourseService.getAll();
+      const courseIds = courses
+        .filter((course) => course.instructor_id === req.user.sub)
+        .map((course) => course.id);
+      const allEnrollments = await EnrollmentService.getAll();
+      const data = allEnrollments.filter((enrollment) => courseIds.includes(enrollment.course_id));
+      return sendSuccess(res, 200, data);
+    }
+
+    return sendSuccess(res, 200, []);
   } catch (err) { next(err); }
 };
 
@@ -54,6 +70,21 @@ exports.patch = async (req, res, next) => {
 
 exports.remove = async (req, res, next) => {
   try {
+    const enrollment = await EnrollmentService.getById(req.params.id);
+    if (!enrollment) return sendError(res, 404, 'Enrollment not found');
+
+    if (req.user.role === 'instructor') {
+      return sendError(res, 403, 'Instructors cannot remove enrollments');
+    }
+
+    if (req.user.role === 'learner' && enrollment.user_id !== req.user.sub) {
+      return sendError(res, 403, 'You can only leave your own courses');
+    }
+
+    if (enrollment.status === 'completed') {
+      return sendError(res, 400, 'Completed courses cannot be left');
+    }
+
     await EnrollmentService.remove(req.params.id);
     sendSuccess(res, 200, null, 'Enrollment deleted');
   } catch (err) { next(err); }
